Readme for adding resources
