+++ src/connectors/huggingface/huggingface_dataset_connector.py
@@ -9,7 +9,8 @@ from connectors import ResourceConnector
 from connectors.resource_with_relations import ResourceWithRelations
 from database.model.dataset import Dataset
 from database.model.dataset.data_download import DataDownloadORM
-from database.model.general import License, Keyword
+from database.model.general.license import License
+from database.model.general.keyword import Keyword
 from platform_names import PlatformName
