+++ src/connectors/openml/openml_dataset_connector.py
@@ -12,7 +12,8 @@ from fastapi import HTTPException
 from connectors.abstract.resource_connector import ResourceConnector
 from database.model.dataset import Dataset
 from database.model.dataset.data_download import DataDownloadORM
-from database.model.general import Keyword, License
+from database.model.general.keyword import Keyword
+from database.model.general.license import License
 from platform_names import PlatformName
